rose slv: new CLI. Support view and update modes.
In update mode, arguments can now be a `*` (for all task jobs), a cycle
time or a task ID.

Support a `--tidy-remote` option to remove job logs on remote hosts
after their retrieval.

Support a `--archive` option (and removed
`--log-archive-threshold=CYCLE`) to switch on archive mode on the
specified cycle times in the argument list.

Switch off view mode by default in update mode, but can be turned on
explicitly with `--view`.

Some internal logic improvements.