rose_prune: user interface refresh.
Functionalities now divided into 4 settings:
* prune-remote-logs-at=cycle ...
* archive-logs-at=cycle ...
* prune-work-at=cycle[:globs] ...
* prune-datac-at=cycle[:globs] ...

The 1st two functionalities call the underlying libraries of `rose
suite-log` to re-sync the remote job logs, prune them from the remote
hosts, (and archive the cycle job logs).

The last two functionalities are to prune items in the work directories
and the share cycle data directories. Globs can be specified for each
cycle so that only matched items in the relevant directories are
pruned.