rosie copy: use single commit
Modify `rosie create` and `rosa svn-pre-commit` to allow a suite copy to
be a single commit.

Add `rosa svn-pre-commit` test battery

Move `rosa-svn-pre-commit` logic into lib/python/rosie/svn_pre_commit.py

Fix `rosa svn-pre-commit` not failing when an author not in the access
list is adding items to the suite trunk.

Add tests for `rosie id` to ensure deleted suites do not cause a
miscount in latest and next.