rose metadata: change option compulsory=true meaning.
This change alters the metadata specification to
allow compulsory options to be missing from a
configuration if and only if their parent section
is also missing.

The compulsory macro logic has been changed to
reflect this and as part of a general refactor
of that logic, which did not handle duplicate
section compulsory option addition properly in
rose edit mode, and which had the --fix bug
described in #1158.