rose edit: fix change meta or project flag.
This fixes changing the =meta or =project variable
within a rose edit session. Due to the old code's
reliance on ConfigNode data, which is not updated as
often, the update of the metadata does not get
triggered. This change adds the ability to source
the meta or project flag from the variable itself,
which is more correct.