rose {app,suite}-run: fix file install performance
An application configuration or suite configuration with large number of free
files can currently take a long time to install, because each file is being
accessed individually. This fix collects them together as top level items.
While this loses some flexibility in fine grain control of the installation
behaviour, the speed up is well worth it.