#1334: rose macro, app-upgrade: handle opt confs.
We can now handle validation and macro transformation of
optional configurations in rose macro and rose app-upgrade.

Validation is easy.

Transforms are hard. We first transform/upgrade the main configuration
by itself to use as a reference point. We then loop through each
optional configuration, combine it with the old main configuration,
and transform/upgrade that. Each optional configuration is then
re-created as a diff of the new transformed object vs the reference
point.