rose_prune: can now prune any cycling items
New `prune{ITEM}=CYCLE[:GLOBS] ...` allow prune of items under any
sub-directories. The old `prune-work-at` and `prune-datac-at` are now
deprecated aliases of `prune{work}` and `prune{sharecycle}`.

Cycle points can now be date-time points or offsets of the current cycle
point.

Each glob in the GLOBS string can now contain a `%(cycle)s`
substitution. When a glob is specified like so, the program will not add
the cycle under ITEM as a sub-directory, but will substitute `%(cycle)s`
in the glob with the cycle.

Now fail if a cycle point in the configuration has a bad syntax.