rose.checksum: new function to guess alogorithm
On file installation for application, suite, etc, if a specified
checksum is calculated with a different algorithm, the system will now
make an attempt to guess the algorithm before comparing the specified
checksum with the actual checksum. For example, if the default checksum
algorithm is MD5, but the expected checksum is calculated using the SHA1
algorithm, the system will now re-calculate the target's checksum using
the SHA1 algorithm before making a comparison.