t: improve robustness
Retry delays in cylc `global.rc` can cause some tests to fail. This
change allows the tests to run in purer form without cylc `global.rc`
settings.
This change also ensures that the new rose_bunch test is able to tidy up
the suite. The NAME variable (suite name) was modified before it is
supplied to the `rose suite-clean` command.