Added --validate-suite-only option to rose suite-run
The --validate-suite-only option bypasses the remote and local installation
of a suite. Instead it will create a temporary copy and only perform the
'cylc validate' command, cleaning up the temporary folder afterwards. No
directories or files are updated in the '~/cylc-run' area on disk.

The benefit of this option is that it allows people to have a simple test
through a CI framework on any pull request, or prior to packaging a system
up for automated deployment. At bare minimum, it is nice to ensure that
a change has not broken the suite.rc setup without any 'permanent' changes
to a disk.